Title: 구간의 합이 같게 나누기, Time: 225ms, Memory: 18MB, Status: Passed - Codetree

+++ 250506/구간의 합이 같게 나누기/divide-by-equal-sum-of-intervals.java
@@ -18,7 +18,7 @@ public class Main {
     static int n;
     static int m;
     static int k;
-    static int ans;
+    static long ans;
     static long [] arr;
     static boolean[][] visit;
     static int[] L;
